refactor: organize config tests into separate files by feature
